backend/pyproject.toml
[project]
name = "calculadora-preco-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "pyjwt[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]
